% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scripps_data_co2}
\alias{scripps_data_co2}
\title{CO2 daily flask data from Scripps stations}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12868 rows and 8 columns.}
\usage{
scripps_data_co2
}
\description{
CO2 flask data of all active Scripps stations, downloaded on Aug 15 2018.
}
\keyword{data}
