% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scripps_stations}
\alias{scripps_stations}
\title{Station information for the Scripps stations}
\format{A data frame with 10 rows and 6 variables:
\describe{
  \item{Station Name}{name of the station and country/state.}
  \item{Station Code}{three letter station abbreviation. }
  \item{Latitude}{geographic latitude in degrees North.}
  \item{longitude}{geographic longitude in degrees East.}
  \item{Elevation}{elevation above sea level in meters.}
  \item{Dates}{Period of active status.}
}}
\usage{
scripps_stations
}
\description{
Information on all Scripps stations active on Aug 15 2018.
}
\keyword{data}
